pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-gitmsg"
version = "0.1.0"
description = "自动生成 Git Commit Message 的命令行工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["git", "commit", "message", "ai", "llm", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
gitmsg = "auto_gitmsg.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/auto-gitmsg"
Repository = "https://github.com/yourusername/auto-gitmsg"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_gitmsg*"]
